refactor: abstract filter states to a context
